requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
huggingface_hub>=0.17.0
pyyaml>=6.0
tqdm>=4.66.0
